Experiments/requirements.txt
torchvision
tqdm
Pillow